+++ merchant-adapters/src/main/java/com/paypulse/merchant/adapters/MerchantRepositoryAdapter.java
@@ -1,14 +1,16 @@
 package com.paypulse.merchant.adapters;
 
+import java.util.Optional;
+import java.util.UUID;
+
 import com.paypulse.merchant.adapters.mapper.MerchantPersistenceMapper;
 import com.paypulse.merchant.adapters.out.MerchantJpaEntity;
 import com.paypulse.merchant.domain.entity.Merchant;
 import com.paypulse.merchant.ports.out.merchant.MerchantPersistencePort;
+
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
-
-import java.util.Optional;
 
 @ApplicationScoped
 public class MerchantRepositoryAdapter implements MerchantPersistencePort {
@@ -16,38 +18,33 @@ public class MerchantRepositoryAdapter implements MerchantPersistencePort {
     @Inject
     MerchantPersistenceMapper mapper;
 
-    /**
-     * @param merchant
-     * @return
-     */
     @Override
     @Transactional
-    public Merchant save(Merchant merchant) {
-        MerchantJpaEntity entity = mapper.toEntity(merchant);
-        entity.persist();
-        return mapper.toDomain(entity);
+    public void save(Merchant merchant) {
+        mapper.toEntity(merchant).persist();
     }
 
-    /**
-     * @param merchantId
-     * @return
-     */
     @Override
     @Transactional
-    public Optional<Merchant> findByMerchantId(String merchantId) {
-        return MerchantJpaEntity
-                .find("merchantId", merchantId)
+    public Optional<Merchant> findByMerchantId(UUID merchantId) {
+        return MerchantJpaEntity.find("merchantId", merchantId)
                 .firstResultOptional()
-                .map(entity ->mapper.toDomain((MerchantJpaEntity) entity));
+                .map(entity -> mapper.toDomain((MerchantJpaEntity) entity));
     }
 
-    /**
-     * @param email
-     * @return
-     */
     @Override
     @Transactional
-    public boolean existsByEmail(String email) {
-        return MerchantJpaEntity.count("merchantEmail", email) > 0;
+    public Optional<Merchant> findByMerchantName(String merchantName) {
+        return MerchantJpaEntity.find("merchantName", merchantName)
+                .firstResultOptional()
+                .map(entity -> mapper.toDomain((MerchantJpaEntity) entity));
+    }
+
+    @Override
+    @Transactional
+    public Optional<Merchant> findByNameCaseInsensitive(String merchantName) {
+        return MerchantJpaEntity.find("lower(merchantName) = lower(?1)", merchantName)
+                .firstResultOptional()
+                .map(entity -> mapper.toDomain((MerchantJpaEntity) entity));
     }
 }