+++ merchant-ports/src/main/java/com/paypulse/merchant/ports/in/merchant/RegisterMerchantUseCase.java
@@ -1,9 +1,9 @@
 package com.paypulse.merchant.ports.in.merchant;
 
 import com.paypulse.merchant.domain.entity.Merchant;
+import com.paypulse.merchant.model.command.RegisterMerchantCommand;
 
 public interface RegisterMerchantUseCase {
 
-    Merchant registerMerchant(Merchant merchant);
-
+    Merchant registerMerchant(RegisterMerchantCommand command);
 }